back to prev 1

+++ train_gpt.py
@@ -367,7 +367,7 @@ class GPT(nn.Module):
         assert num_layers % 2 == 0
         #self.skip_weights = nn.Parameter(torch.ones(num_layers // 2))
         self.residual_weights = nn.Parameter(torch.ones(num_layers, 1, dtype=torch.bfloat16))
-        self.x0_weight = nn.Parameter(torch.tensor([0.0]))
+        self.x0_weight = nn.Parameter(torch.tensor([0.0], dtype=torch.bfloat16))
         #fan_in = num_layers // 2
         #std = 1 / math.sqrt(fan_in)  # Standard deviation
         #nn.init.normal_(self.skip_weights, mean=0.0, std=std)
